Split GenericValue into protocol specific modules
