+++ Cargo.toml
@@ -28,7 +28,7 @@ codegen-units = 1
 
 [workspace.dependencies]
 scylla = { version = "0.11.0", features = ["ssl"] }
-bytes = "1.0.0"
+bytes = { version = "1.0.0", features = ["serde"] }
 tokio = { version = "1.25.0", features = ["full", "macros"] }
 tokio-util = { version = "0.7.7" }
 tokio-openssl = "0.6.2"
